+++ server/src/test/scala/cromwell/SimpleWorkflowActorSpec.scala
@@ -69,7 +69,8 @@ class SimpleWorkflowActorSpec extends CromwellTestKitWordSpec with BeforeAndAfte
         workflowStoreActor = system.actorOf(Props.empty),
         workflowHeartbeatConfig = WorkflowHeartbeatConfig(config),
         totalJobsByRootWf = new AtomicInteger(),
-        fileHashCacheActor = None),
+        fileHashCacheActor = None,
+        blacklistCache = None),
       supervisor = supervisor.ref,
       name = s"workflow-actor-$workflowId"
     )