+++ supportedBackends/google/pipelines/common/src/main/scala/cromwell/backend/google/pipelines/common/api/PipelinesApiRequestFactory.scala
@@ -8,6 +8,7 @@ import cromwell.backend.standard.StandardAsyncJob
 import cromwell.core.labels.Labels
 import cromwell.core.logging.JobLogger
 import cromwell.core.path.Path
+import wom.runtime.WomOutputRuntimeExtractor
 
 /**
   * The PipelinesApiRequestFactory defines the HttpRequests needed to run jobs
@@ -71,7 +72,8 @@ object PipelinesApiRequestFactory {
                                       labels: Labels,
                                       preemptible: Boolean,
                                       jobShell: String,
-                                      privateDockerKeyAndEncryptedToken: Option[CreatePipelineDockerKeyAndToken]) {
+                                      privateDockerKeyAndEncryptedToken: Option[CreatePipelineDockerKeyAndToken],
+                                      womOutputRuntimeExtractor: Option[WomOutputRuntimeExtractor]) {
     def literalInputs = inputOutputParameters.literalInputParameters
     def inputParameters = inputOutputParameters.fileInputParameters
     def outputParameters = inputOutputParameters.fileOutputParameters