+++ supportedBackends/google/pipelines/v2alpha1/src/main/scala/cromwell/backend/google/pipelines/v2alpha1/PipelinesUtilityConversions.scala
@@ -10,6 +10,9 @@ import cromwell.core.ExecutionEvent
 import cromwell.core.logging.JobLogger
 import mouse.all._
 
+import scala.language.postfixOps
+import scala.util.Try
+
 trait PipelinesUtilityConversions {
   def toAccelerator(gpuResource: GpuResource) = new Accelerator().setCount(gpuResource.gpuCount.value.toLong).setType(gpuResource.gpuType.toString)
   def toMachineType(jobLogger: JobLogger)(attributes: PipelinesApiRuntimeAttributes) = MachineConstraints.machineType(attributes.memory, attributes.cpu, jobLogger)
@@ -22,7 +25,26 @@ trait PipelinesUtilityConversions {
     .setName(disk.name)
     .setSizeGb(disk.sizeGb)
     .setType(disk.diskType |> toV2DiskType)
-  def toExecutionEvent(event: Event) : ExecutionEvent = ExecutionEvent(event.getDescription, OffsetDateTime.parse(event.getTimestamp))
+
+  def toExecutionEvent(actionIndexToEventType: Map[Int, String])(event: Event): ExecutionEvent = {
+    val groupingFromAction = for {
+      rawValue <- Option(event.getDetails.get("actionId"))
+      integerValue <- Try(Integer.valueOf(rawValue.toString)).toOption
+      group <- actionIndexToEventType.get(integerValue)
+    } yield group
+
+    // There are both "Started pulling" and "Stopped pulling" events but these are confusing for metadata, especially on the
+    // timing diagram. Create a single "Pulling <docker image>" grouping to absorb these events.
+    def groupingFromPull: Option[String] = List("Started", "Stopped") flatMap { k =>
+      Option(event.getDescription) collect { case d if d.startsWith(s"$k pulling") => "Pulling" + d.substring(s"$k pulling".length)}
+    } headOption
+
+    ExecutionEvent(
+      name = event.getDescription,
+      offsetDateTime = OffsetDateTime.parse(event.getTimestamp),
+      grouping = groupingFromAction.orElse(groupingFromPull)
+    )
+  }
 
   private def toV2DiskType(diskType: DiskType) = diskType match {
     case DiskType.HDD => "pd-standard"