+++ supportedBackends/google/pipelines/v2alpha1/src/main/scala/cromwell/backend/google/pipelines/v2alpha1/api/Delocalization.scala
@@ -8,12 +8,15 @@ import common.util.StringUtil._
 import cromwell.backend.google.pipelines.common.PipelinesApiAttributes.LocalizationConfiguration
 import cromwell.backend.google.pipelines.common.api.PipelinesApiRequestFactory.CreatePipelineParameters
 import cromwell.backend.google.pipelines.v2alpha1.PipelinesConversions._
+import cromwell.backend.google.pipelines.v2alpha1.RuntimeOutputMapping
 import cromwell.backend.google.pipelines.v2alpha1.ToParameter.ops._
-import cromwell.backend.google.pipelines.v2alpha1.api.ActionBuilder.Labels.Value
+import cromwell.backend.google.pipelines.v2alpha1.api.ActionBuilder.Labels.{Key, Value}
 import cromwell.backend.google.pipelines.v2alpha1.api.ActionBuilder._
 import cromwell.backend.google.pipelines.v2alpha1.api.ActionCommands._
 import cromwell.backend.google.pipelines.v2alpha1.api.Delocalization._
 import cromwell.core.path.{DefaultPathBuilder, Path}
+import org.apache.commons.codec.binary.Base64
+import wom.runtime.WomOutputRuntimeExtractor
 
 import scala.collection.JavaConverters._
 import scala.concurrent.duration._
@@ -30,62 +33,88 @@ trait Delocalization {
   private def delocalizeLogsAction(gcsLogPath: Path)(implicit localizationConfiguration: LocalizationConfiguration) = {
     cloudSdkShellAction(
     delocalizeDirectory(DefaultPathBuilder.build(logsRoot).get, gcsLogPath, plainTextContentType)
-    )(flags = List(ActionFlag.AlwaysRun))
+    )(flags = List(ActionFlag.AlwaysRun), labels = Map(Key.Tag -> Value.Delocalization))
   }
 
   // Used for the final copy of the logs to make sure we have the most up to date version before terminating the job
   private def copyAggregatedLogToLegacyPath(callExecutionContainerRoot: Path, gcsLegacyLogPath: Path)(implicit localizationConfiguration: LocalizationConfiguration): Action = {
     cloudSdkShellAction(
       delocalizeFileTo(DefaultPathBuilder.build(aggregatedLog).get, gcsLegacyLogPath, plainTextContentType)
-    )(flags = List(ActionFlag.AlwaysRun))
+    )(flags = List(ActionFlag.AlwaysRun), labels = Map(Key.Tag -> Value.Delocalization))
   }
 
   // Periodically copies the logs out to GCS
   private def copyAggregatedLogToLegacyPathPeriodic(callExecutionContainerRoot: Path, gcsLegacyLogPath: Path)(implicit localizationConfiguration: LocalizationConfiguration): Action = {
     cloudSdkShellAction(
       every(30.seconds) { delocalizeFileTo(DefaultPathBuilder.build(aggregatedLog).get, gcsLegacyLogPath, plainTextContentType) }
-    )(flags = List(ActionFlag.RunInBackground))
+    )(flags = List(ActionFlag.RunInBackground), labels = Map(Key.Tag -> Value.Background))
   }
 
-  private def parseOutputJsonAction(containerCallRoot: String, outputDirectory: String, outputFile: String, mounts: List[Mount]): Action = {
+  private def runtimeOutputExtractorAction(containerCallRoot: String,
+                                    outputFile: String,
+                                    mounts: List[Mount],
+                                    womOutputRuntimeExtractor: WomOutputRuntimeExtractor): Action = {
     val commands = List(
       "-c",
-      // Create the directory where the output of jq will be written
-      s"mkdir -p $outputDirectory &&" +
-      // Read the content of cwl.output.json - redirect stderr to /dev/null so it doesn't fail if there's no cwl.output.json
-      s" cat ${containerCallRoot.ensureSlashed}cwl.output.json 2>/dev/null" +
-        // Pipe the result to jq. Parse the json and traverse it looking for "path" or "location" values,
-        // stripping any fully qualified file:// URLs.
-      """ | jq -r '.. | .path? // .location? // empty | gsub("file://"; "")' """ +
-      // Redirect the result to a file that can be used in the next action to delocalize the extracted files / directories
-      s" > $outputFile"
+      // Create the directory where the fofn will be written
+      s"mkdir -p $$(dirname $outputFile) && " +
+      s"cd $containerCallRoot && " +
+      s"${womOutputRuntimeExtractor.command} > $outputFile"
     )
 
     ActionBuilder
-      .withImage("stedolan/jq@sha256:a61ed0bca213081b64be94c5e1b402ea58bc549f457c2682a86704dd55231e09")
+      .withImage(womOutputRuntimeExtractor.dockerImage.getOrElse(cloudSdkImage))
       .setCommands(commands.asJava)
       .withMounts(mounts)
       .setEntrypoint("/bin/bash")
       // Saves us some time if something else fails before we get to run this action
       .withFlags(List(ActionFlag.DisableImagePrefetch))
+      .withLabels(Map(Key.Tag -> Value.Delocalization))
+  }
+  
+  private def delocalizeRuntimeOutputsScript(fofnPath: String, workflowRoot: Path, cloudCallRoot: Path) = {
+    val gsutilCommand: String => String = { flag =>
+      s"""gsutil -m $flag cp -r $$line "${cloudCallRoot.pathAsString.ensureSlashed}$$gcs_path""""
+    }
+    
+    def sedStripPrefix(prefix: String) = s"""sed -e "s/^${prefix.ensureSedEscaped}//""""
+    
+    // See RuntimeOutputMapping.prefixFilters for more info on why this is needed
+    val prefixFilters = RuntimeOutputMapping
+      .prefixFilters(workflowRoot)
+      .map(sedStripPrefix)
+      .mkString(" | ")
+
+    /*
+     * Delocalize all the files returned by the runtime output extractor
+     */
+    s"""|#!/bin/bash
+        |
+        |set -x
+        |
+        |if [ -f $fofnPath ]; then
+        |  while IFS= read line
+        |  do
+        |    gcs_path=$$(echo $$line | $prefixFilters)
+        |    ${recoverRequesterPaysError(cloudCallRoot)(gsutilCommand)}
+        |  done  <$fofnPath
+        |fi""".stripMargin
   }
 
-  private def delocalizeOutputJsonFilesAction(cloudCallRoot: Path, inputFile: String, workflowRoot: String, mounts: List[Mount]): Action = {
-    val sedStripSlashPrefix = "s/^\\///"
-    val gsutilCommand: String => String = flag => s"""gsutil -m $flag cp -r % ${cloudCallRoot.pathAsString.ensureSlashed}$$(echo % | sed -e "$sedStripSlashPrefix")"""
-    val command =
-        // Read the file containing files to delocalize
-        s"cat $inputFile 2>/dev/null" +
-        /*
-         * Pipe the result to xargs and execute a gsutil cp command, appending the path to the cloudCallRoot
-         * Use a subshell so that we can strip the potential leading slash from the local path and avoid double slashes in the cloud path
-         * It also seems that the gsutil fails without sub shelling
-         * We can't use the gsutil -I flag here because it would lose the directory structure once it gets copied to the bucket
-         * sh -c 'gsutil cp % $(echo % | sed -e "s/^\///")'
-         */
-        s""" | xargs -I % sh -c '${recoverRequesterPaysError(cloudCallRoot)(gsutilCommand)}'"""
+  private def delocalizeRuntimeOutputsAction(cloudCallRoot: Path, inputFile: String, workflowRoot: Path, mounts: List[Mount]): Action = {
+    def multiLineCommand(commandString: String) = {
+      val randomUuid = UUID.randomUUID().toString
+      val base64EncodedScript = Base64.encodeBase64String(commandString.getBytes)
+      val scriptPath = s"/tmp/$randomUuid.sh"
 
-    ActionBuilder.cloudSdkShellAction(command)(mounts)
+      s"apt-get install --assume-yes coreutils && " +
+        s"echo $base64EncodedScript | base64 --decode > $scriptPath && " +
+        s"chmod u+x $scriptPath && " +
+        s"sh $scriptPath"
+    }
+    
+    val command = multiLineCommand(delocalizeRuntimeOutputsScript(inputFile, workflowRoot, cloudCallRoot))
+    ActionBuilder.cloudSdkShellAction(command)(mounts, flags = List(ActionFlag.DisableImagePrefetch), labels = Map(Key.Tag -> Value.Delocalization))
   }
 
   def deLocalizeActions(createPipelineParameters: CreatePipelineParameters,
@@ -103,12 +132,17 @@ trait Delocalization {
      */
     val temporaryFofnDirectoryForCwlOutputJson = callExecutionContainerRoot.pathAsString.ensureSlashed + UUID.randomUUID().toString.split("-")(0)
     val temporaryFofnForCwlOutputJson = temporaryFofnDirectoryForCwlOutputJson + "/cwl_output_json_references.txt"
-    val parseAction = parseOutputJsonAction(callExecutionContainerRoot.pathAsString, temporaryFofnDirectoryForCwlOutputJson, temporaryFofnForCwlOutputJson, mounts)
-    val delocalizeAction = delocalizeOutputJsonFilesAction(cloudCallRoot, temporaryFofnForCwlOutputJson, createPipelineParameters.cloudWorkflowRoot.pathAsString, mounts)
+
+    val runtimeExtractionActions = createPipelineParameters.womOutputRuntimeExtractor.toList flatMap { extractor =>
+      List (
+        runtimeOutputExtractorAction(callExecutionContainerRoot.pathAsString, temporaryFofnForCwlOutputJson, mounts, extractor),
+        delocalizeRuntimeOutputsAction(cloudCallRoot, temporaryFofnForCwlOutputJson, createPipelineParameters.cloudWorkflowRoot, mounts)
+      )
+    }
 
     ActionBuilder.annotateTimestampedActions("delocalization", Value.Delocalization)(
       createPipelineParameters.outputParameters.flatMap(_.toActions(mounts).toList) ++
-        List(parseAction, delocalizeAction)
+        runtimeExtractionActions
     ) :+
       copyAggregatedLogToLegacyPath(callExecutionContainerRoot, gcsLegacyLogPath) :+
       copyAggregatedLogToLegacyPathPeriodic(callExecutionContainerRoot, createPipelineParameters.logGcsPath) :+