+++ CHANGELOG.md
@@ -2,10 +2,18 @@
 
 ## 37 Release Notes
 
+### Call cache blacklisting
+
+The Google Pipelines API (PAPI) version 1 and 2 backends now offer the option of call cache blacklisting on a per-bucket basis.
+More info [here](http://cromwell.readthedocs.io/en/develop/CallCaching/#call-cache-copy-authorization-failure-prefix-blacklisting).
+
 ### Changing configuration options
 
 When the value `exit-code-timeout-seconds` is set, `check-alive` command is now only called once every timeout interval instead of each poll.
 
+### Bug fixes
+
+- Fixed a regression in Cromwell 36 that could cause operations on empty arrays to fail with a spurious type error (closes [#4318](https://github.com/broadinstitute/cromwell/issues/4318))
 
 ## 36 Release Notes